+++ packages/core/styles/src/packages/composite/modal/index.ts
@@ -1,8 +1,19 @@
-import { styles, modalSprinkle } from "./nimbus-modal.css";
+import {
+  styles,
+  modalSprinkle,
+  modalHeaderSprinkle,
+  modalBodySprinkle,
+  modalFooterSprinkle,
+} from "./nimbus-modal.css";
 
 export const modal = {
-  classnames: { ...styles },
   ...modalSprinkle,
+  classnames: { ...styles },
+  subComponents: {
+    header: modalHeaderSprinkle,
+    body: modalBodySprinkle,
+    footer: modalFooterSprinkle,
+  },
 };
 
 export type { ModalSprinkle } from "./nimbus-modal.types";