+++ packages/react/src/composite/Card/CHANGELOG.md
@@ -6,8 +6,8 @@ The Card component allows us to group contents and related actions, making it ea
 
 #### 🎉 New features
 
-- Added `Padding` property to the Component `Card.Header`. ([#168](https://github.com/TiendaNube/nimbus-design-system/pull/168) by [@juniorconquista](https://github.com/juniorconquista))
-- Added `Padding` property to the Component `Card.Footer`. ([#168](https://github.com/TiendaNube/nimbus-design-system/pull/168) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `Padding` property to the Component `Card.Header`. ([#169](https://github.com/TiendaNube/nimbus-design-system/pull/169) by [@juniorconquista](https://github.com/juniorconquista))
+- Added `Padding` property to the Component `Card.Footer`. ([#169](https://github.com/TiendaNube/nimbus-design-system/pull/169) by [@juniorconquista](https://github.com/juniorconquista))
 
 ## 2023-04-13 `3.0.1`